Time: 0 ms (100%), Space: 7.8 MB (62.21%) - LeetHub

+++ 0693-binary-number-with-alternating-bits/0693-binary-number-with-alternating-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool hasAlternatingBits(int n) {
-        long long int  ans = n ^ (n >> 1);
+        unsigned int  ans = n ^ (n >> 1);
         return (ans > 0 && (ans & (ans + 1)) == 0);
     }
 };
